requirements.txt
PyJWT
discord.py
pyrebase
python-dotenv
humanfriendly
requests
genshin
pytz
translate
requests